+++ mapgen/z.c
@@ -124,7 +124,7 @@ str_save(char *s)
 {
 	char *p;
 
-	p = my_malloc(strlen(s) + 1);
+	p = my_malloc((unsigned) (strlen(s) + 1));	/* string len fits 32 bits */
 	strcpy(p, s);
 
 	return p;
@@ -256,7 +256,7 @@ getlin_ew(FILE *fp)
 #define MAX_BUF         8192
 
 static char linebuf[MAX_BUF];
-static int nread;
+static ssize_t nread;	/* read() result; ssize_t avoids LP64 truncation */
 static int line_fd = -1;
 static char *point;
 
@@ -533,8 +533,8 @@ fuzzy_one_bad(char *one, char *two, int l1, int l2)
 int
 fuzzy_strcmp(char *one, char *two)
 {
-	int l1 = strlen(one);
-	int l2 = strlen(two);
+	int l1 = (int) strlen(one);	/* name lengths fit 32 bits */
+	int l2 = (int) strlen(two);
 
 	if (l2 >= 4 && fuzzy_transpose(one, two, l1, l2))
 		return TRUE;
